+++ src/main/java/com/telepathicgrunt/world_blender/features/WBConfiguredFeatures.java
@@ -11,15 +11,13 @@ import net.minecraft.world.gen.feature.FeatureConfig;
 public class WBConfiguredFeatures
 {
     public static final ConfiguredFeature<?,?> WB_PORTAL_ALTAR = WBFeatures.WB_PORTAL_ALTAR.configure(FeatureConfig.DEFAULT);
-    public static final ConfiguredFeature<?,?> NO_FLOATING_LIQUIDS_OR_FALLING_BLOCKS = WBFeatures.NO_FLOATING_LIQUIDS_OR_FALLING_BLOCKS.configure(FeatureConfig.DEFAULT);
-    public static final ConfiguredFeature<?,?> SEPARATE_LAVA_AND_WATER = WBFeatures.SEPARATE_LAVA_AND_WATER.configure(FeatureConfig.DEFAULT);
-    
+    public static final ConfiguredFeature<?,?> ANTI_FLOATING_BLOCKS_AND_SEPARATE_LIQUIDS = WBFeatures.ANTI_FLOATING_BLOCKS_AND_SEPARATE_LIQUIDS.configure(FeatureConfig.DEFAULT);
+
     public static void registerConfiguredFeatures()
     {
         MutableRegistry<ConfiguredFeature<?, ?>> registry = (MutableRegistry<ConfiguredFeature<?, ?>>) BuiltinRegistries.CONFIGURED_FEATURE;
 
         Registry.register(registry, new Identifier(WorldBlender.MODID, "portal_altar"), WB_PORTAL_ALTAR);
-        Registry.register(registry, new Identifier(WorldBlender.MODID, "no_floating_liquids_or_falling_blocks"), NO_FLOATING_LIQUIDS_OR_FALLING_BLOCKS);
-        Registry.register(registry, new Identifier(WorldBlender.MODID, "separate_lava_and_water"), SEPARATE_LAVA_AND_WATER);
+        Registry.register(registry, new Identifier(WorldBlender.MODID, "anti_floating_blocks_and_separate_liquids"), ANTI_FLOATING_BLOCKS_AND_SEPARATE_LIQUIDS);
     }
 }